requirements.txt
huggingface_hub
datasets
numpy
torch
torch.utils.data
torchvision
scikit-learn
torchvision
matplotlib.pyplot